requirements.txt
Flask
werkzeug
sqlalchemy
pytest
jsonschema
pycryptodomex